+++ src/interface/export/export_xml.cc
@@ -24,35 +24,35 @@
 
 BEGIN_YAFARAY
 
-XmlExport::XmlExport(const char *fname, const ::yafaray_LoggerCallback_t logger_callback, void *callback_data, ::yafaray_DisplayConsole_t logger_display_console) : Interface(logger_callback, callback_data, logger_display_console), xml_name_(std::string(fname))
+ExportXml::ExportXml(const char *fname, const ::yafaray_LoggerCallback_t logger_callback, void *callback_data, ::yafaray_DisplayConsole_t logger_display_console) : Interface(logger_callback, callback_data, logger_display_console), file_name_(std::string(fname))
 {
-	xml_file_.open(xml_name_.c_str());
-	if(!xml_file_.is_open())
+	file_.open(file_name_.c_str());
+	if(!file_.is_open())
 	{
-		logger_->logError("XmlExport: Couldn't open ", xml_name_);
+		logger_->logError("XmlExport: Couldn't open ", file_name_);
 		return;
 	}
-	else logger_->logInfo("XmlExport: Writing scene to: ", xml_name_);
-	xml_file_ << std::boolalpha;
-	xml_file_ << "<?xml version=\"1.0\"?>\n";
+	else logger_->logInfo("XmlExport: Writing scene to: ", file_name_);
+	file_ << std::boolalpha;
+	file_ << "<?xml version=\"1.0\"?>\n";
 }
 
-void XmlExport::createScene() noexcept
+void ExportXml::createScene() noexcept
 {
-	xml_file_ << "<scene>\n\n";
-	xml_file_ << "<scene_parameters>\n";
+	file_ << "<scene>\n\n";
+	file_ << "<scene_parameters>\n";
 	writeParamMap(*params_);
 	params_->clear();
-	xml_file_ << "</scene_parameters>\n";
+	file_ << "</scene_parameters>\n";
 }
 
-void XmlExport::clearAll() noexcept
+void ExportXml::clearAll() noexcept
 {
 	if(logger_->isVerbose()) logger_->logVerbose("XmlExport: cleaning up...");
-	if(xml_file_.is_open())
+	if(file_.is_open())
 	{
-		xml_file_.flush();
-		xml_file_.close();
+		file_.flush();
+		file_.close();
 	}
 	params_->clear();
 	eparams_->clear();
@@ -60,135 +60,135 @@ void XmlExport::clearAll() noexcept
 	next_obj_ = 0;
 }
 
-void XmlExport::defineLayer() noexcept
+void ExportXml::defineLayer() noexcept
 {
-	xml_file_ << "\n<layer>\n";
+	file_ << "\n<layer>\n";
 	writeParamMap(*params_);
 	params_->clear();
-	xml_file_ << "</layer>\n";
+	file_ << "</layer>\n";
 }
 
-bool XmlExport::startGeometry() noexcept { return true; }
+bool ExportXml::startGeometry() noexcept { return true; }
 
-bool XmlExport::endGeometry() noexcept { return true; }
+bool ExportXml::endGeometry() noexcept { return true; }
 
-unsigned int XmlExport::getNextFreeId() noexcept
+unsigned int ExportXml::getNextFreeId() noexcept
 {
 	return ++next_obj_;
 }
 
-bool XmlExport::endObject() noexcept
+bool ExportXml::endObject() noexcept
 {
-	xml_file_ << "</object>\n";
+	file_ << "</object>\n";
 	return true;
 }
 
-int XmlExport::addVertex(double x, double y, double z) noexcept
+int ExportXml::addVertex(double x, double y, double z) noexcept
 {
-	xml_file_ << "\t<p x=\"" << x << "\" y=\"" << y << "\" z=\"" << z << "\"/>\n";
+	file_ << "\t<p x=\"" << x << "\" y=\"" << y << "\" z=\"" << z << "\"/>\n";
 	return 0;
 }
 
-int XmlExport::addVertex(double x, double y, double z, double ox, double oy, double oz) noexcept
+int ExportXml::addVertex(double x, double y, double z, double ox, double oy, double oz) noexcept
 {
-	xml_file_ << "\t<p x=\"" << x << "\" y=\"" << y << "\" z=\"" << z
+	file_ << "\t<p x=\"" << x << "\" y=\"" << y << "\" z=\"" << z
 			  << "\" ox=\"" << ox << "\" oy=\"" << oy << "\" oz=\"" << oz << "\"/>\n";
 	return 0;
 }
 
-void XmlExport::addNormal(double x, double y, double z) noexcept
+void ExportXml::addNormal(double x, double y, double z) noexcept
 {
-	xml_file_ << "\t<n x=\"" << x << "\" y=\"" << y << "\" z=\"" << z << "\"/>\n";
+	file_ << "\t<n x=\"" << x << "\" y=\"" << y << "\" z=\"" << z << "\"/>\n";
 }
 
-void XmlExport::setCurrentMaterial(const char *name) noexcept
+void ExportXml::setCurrentMaterial(const char *name) noexcept
 {
 	const std::string name_str(name);
 	if(name_str != current_material_) //need to set current material
 	{
-		xml_file_ << "\t<set_material sval=\"" << name_str << "\"/>\n";
+		file_ << "\t<set_material sval=\"" << name_str << "\"/>\n";
 		current_material_ = name_str;
 	}
 }
 
-bool XmlExport::addFace(int a, int b, int c) noexcept
+bool ExportXml::addFace(int a, int b, int c) noexcept
 {
-	xml_file_ << "\t<f a=\"" << a << "\" b=\"" << b << "\" c=\"" << c << "\"/>\n";
+	file_ << "\t<f a=\"" << a << "\" b=\"" << b << "\" c=\"" << c << "\"/>\n";
 	return true;
 }
 
-bool XmlExport::addFace(int a, int b, int c, int uv_a, int uv_b, int uv_c) noexcept
+bool ExportXml::addFace(int a, int b, int c, int uv_a, int uv_b, int uv_c) noexcept
 {
-	xml_file_ << "\t<f a=\"" << a << "\" b=\"" << b << "\" c=\"" << c
+	file_ << "\t<f a=\"" << a << "\" b=\"" << b << "\" c=\"" << c
 			  << "\" uv_a=\"" << uv_a << "\" uv_b=\"" << uv_b << "\" uv_c=\"" << uv_c << "\"/>\n";
 	return true;
 }
 
-int XmlExport::addUv(float u, float v) noexcept
+int ExportXml::addUv(float u, float v) noexcept
 {
-	xml_file_ << "\t<uv u=\"" << u << "\" v=\"" << v << "\"/>\n";
+	file_ << "\t<uv u=\"" << u << "\" v=\"" << v << "\"/>\n";
 	return n_uvs_++;
 }
 
-bool XmlExport::smoothMesh(const char *name, double angle) noexcept
+bool ExportXml::smoothMesh(const char *name, double angle) noexcept
 {
-	xml_file_ << "<smooth object_name=\"" << name << "\" angle=\"" << angle << "\"/>\n";
+	file_ << "<smooth object_name=\"" << name << "\" angle=\"" << angle << "\"/>\n";
 	return true;
 }
 
-void XmlExport::writeMatrix(const std::string &name, const Matrix4 &m, std::ofstream &xml_file) noexcept
+void ExportXml::writeMatrix(const std::string &name, const Matrix4 &m, std::ofstream &file) noexcept
 {
-	xml_file << "<" << name << " m00=\"" << m[0][0] << "\" m01=\"" << m[0][1] << "\" m02=\"" << m[0][2] << "\" m03=\"" << m[0][3] << "\""
-			 << " m10=\"" << m[1][0] << "\" m11=\"" << m[1][1] << "\" m12=\"" << m[1][2] << "\" m13=\"" << m[1][3] << "\""
-			 << " m20=\"" << m[2][0] << "\" m21=\"" << m[2][1] << "\" m22=\"" << m[2][2] << "\" m23=\"" << m[2][3] << "\""
-			 << " m30=\"" << m[3][0] << "\" m31=\"" << m[3][1] << "\" m32=\"" << m[3][2] << "\" m33=\"" << m[3][3] << "\"/>";
+	file << "<" << name << " m00=\"" << m[0][0] << "\" m01=\"" << m[0][1] << "\" m02=\"" << m[0][2] << "\" m03=\"" << m[0][3] << "\""
+		 << " m10=\"" << m[1][0] << "\" m11=\"" << m[1][1] << "\" m12=\"" << m[1][2] << "\" m13=\"" << m[1][3] << "\""
+		 << " m20=\"" << m[2][0] << "\" m21=\"" << m[2][1] << "\" m22=\"" << m[2][2] << "\" m23=\"" << m[2][3] << "\""
+		 << " m30=\"" << m[3][0] << "\" m31=\"" << m[3][1] << "\" m32=\"" << m[3][2] << "\" m33=\"" << m[3][3] << "\"/>";
 }
 
-void XmlExport::writeParam(const std::string &name, const Parameter &param, std::ofstream &xml_file, ColorSpace xml_color_space, float xml_gamma) noexcept
+void ExportXml::writeParam(const std::string &name, const Parameter &param, std::ofstream &file, ColorSpace color_space, float gamma) noexcept
 {
 	const Parameter::Type type = param.type();
 	if(type == Parameter::Int)
 	{
 		int i = 0;
 		param.getVal(i);
-		xml_file << "<" << name << " ival=\"" << i << "\"/>\n";
+		file << "<" << name << " ival=\"" << i << "\"/>\n";
 	}
 	else if(type == Parameter::Bool)
 	{
 		bool b = false;
 		param.getVal(b);
-		xml_file << "<" << name << " bval=\"" << b << "\"/>\n";
+		file << "<" << name << " bval=\"" << b << "\"/>\n";
 	}
 	else if(type == Parameter::Float)
 	{
 		double f = 0.0;
 		param.getVal(f);
-		xml_file << "<" << name << " fval=\"" << f << "\"/>\n";
+		file << "<" << name << " fval=\"" << f << "\"/>\n";
 	}
 	else if(type == Parameter::String)
 	{
 		std::string s;
 		param.getVal(s);
-		if(!s.empty()) xml_file << "<" << name << " sval=\"" << s << "\"/>\n";
+		if(!s.empty()) file << "<" << name << " sval=\"" << s << "\"/>\n";
 	}
 	else if(type == Parameter::Vector)
 	{
 		Point3 p(0.f);
 		param.getVal(p);
-		xml_file << "<" << name << " x=\"" << p.x_ << "\" y=\"" << p.y_ << "\" z=\"" << p.z_ << "\"/>\n";
+		file << "<" << name << " x=\"" << p.x_ << "\" y=\"" << p.y_ << "\" z=\"" << p.z_ << "\"/>\n";
 	}
 	else if(type == Parameter::Color)
 	{
 		Rgba c(0.f);
 		param.getVal(c);
-		c.colorSpaceFromLinearRgb(xml_color_space, xml_gamma);    //Color values are encoded to the desired color space before saving them to the XML file
-		xml_file << "<" << name << " r=\"" << c.r_ << "\" g=\"" << c.g_ << "\" b=\"" << c.b_ << "\" a=\"" << c.a_ << "\"/>\n";
+		c.colorSpaceFromLinearRgb(color_space, gamma);    //Color values are encoded to the desired color space before saving them to the XML file
+		file << "<" << name << " r=\"" << c.r_ << "\" g=\"" << c.g_ << "\" b=\"" << c.b_ << "\" a=\"" << c.a_ << "\"/>\n";
 	}
 	else if(type == Parameter::Matrix)
 	{
 		Matrix4 m;
 		param.getVal(m);
-		writeMatrix(name, m, xml_file);
+		writeMatrix(name, m, file);
 	}
 	else
 	{
@@ -196,142 +196,142 @@ void XmlExport::writeParam(const std::string &name, const Parameter &param, std:
 	}
 }
 
-bool XmlExport::addInstance(const char *base_object_name, const Matrix4 &obj_to_world) noexcept
+bool ExportXml::addInstance(const char *base_object_name, const Matrix4 &obj_to_world) noexcept
 {
-	xml_file_ << "\n<instance base_object_name=\"" << base_object_name << "\" >\n\t";
-	writeMatrix("transform", obj_to_world, xml_file_);
-	xml_file_ << "\n</instance>\n";
+	file_ << "\n<instance base_object_name=\"" << base_object_name << "\" >\n\t";
+	writeMatrix("transform", obj_to_world, file_);
+	file_ << "\n</instance>\n";
 	return true;
 }
 
-void XmlExport::writeParamMap(const ParamMap &param_map, int indent) noexcept
+void ExportXml::writeParamMap(const ParamMap &param_map, int indent) noexcept
 {
 	const std::string tabs(indent, '\t');
 	for(const auto &param : param_map)
 	{
-		xml_file_ << tabs;
-		writeParam(param.first, param.second, xml_file_, xml_color_space_, xml_gamma_);
+		file_ << tabs;
+		writeParam(param.first, param.second, file_, color_space_, gamma_);
 	}
 }
 
-void XmlExport::writeParamList(int indent) noexcept
+void ExportXml::writeParamList(int indent) noexcept
 {
 	const std::string tabs(indent, '\t');
 	for(const auto &param : *eparams_)
 	{
-		xml_file_ << tabs << "<list_element>\n";
+		file_ << tabs << "<list_element>\n";
 		writeParamMap(param, indent + 1);
-		xml_file_ << tabs << "</list_element>\n";
+		file_ << tabs << "</list_element>\n";
 	}
 }
 
-Light *XmlExport::createLight(const char *name) noexcept
+Light *ExportXml::createLight(const char *name) noexcept
 {
-	xml_file_ << "\n<light name=\"" << name << "\">\n";
+	file_ << "\n<light name=\"" << name << "\">\n";
 	writeParamMap(*params_);
-	xml_file_ << "</light>\n";
+	file_ << "</light>\n";
 	return nullptr;
 }
 
-Texture *XmlExport::createTexture(const char *name) noexcept
+Texture *ExportXml::createTexture(const char *name) noexcept
 {
-	xml_file_ << "\n<texture name=\"" << name << "\">\n";
+	file_ << "\n<texture name=\"" << name << "\">\n";
 	writeParamMap(*params_);
-	xml_file_ << "</texture>\n";
+	file_ << "</texture>\n";
 	return nullptr;
 }
 
-Material *XmlExport::createMaterial(const char *name) noexcept
+Material *ExportXml::createMaterial(const char *name) noexcept
 {
-	xml_file_ << "\n<material name=\"" << name << "\">\n";
+	file_ << "\n<material name=\"" << name << "\">\n";
 	writeParamMap(*params_);
 	writeParamList(1);
-	xml_file_ << "</material>\n";
+	file_ << "</material>\n";
 	return nullptr;
 }
-Camera *XmlExport::createCamera(const char *name) noexcept
+Camera *ExportXml::createCamera(const char *name) noexcept
 {
-	xml_file_ << "\n<camera name=\"" << name << "\">\n";
+	file_ << "\n<camera name=\"" << name << "\">\n";
 	writeParamMap(*params_);
-	xml_file_ << "</camera>\n";
+	file_ << "</camera>\n";
 	return nullptr;
 }
-Background *XmlExport::createBackground(const char *name) noexcept
+Background *ExportXml::createBackground(const char *name) noexcept
 {
-	xml_file_ << "\n<background name=\"" << name << "\">\n";
+	file_ << "\n<background name=\"" << name << "\">\n";
 	writeParamMap(*params_);
-	xml_file_ << "</background>\n";
+	file_ << "</background>\n";
 	return nullptr;
 }
-Integrator *XmlExport::createIntegrator(const char *name) noexcept
+Integrator *ExportXml::createIntegrator(const char *name) noexcept
 {
-	xml_file_ << "\n<integrator name=\"" << name << "\">\n";
+	file_ << "\n<integrator name=\"" << name << "\">\n";
 	writeParamMap(*params_);
-	xml_file_ << "</integrator>\n";
+	file_ << "</integrator>\n";
 	return nullptr;
 }
 
-VolumeRegion *XmlExport::createVolumeRegion(const char *name) noexcept
+VolumeRegion *ExportXml::createVolumeRegion(const char *name) noexcept
 {
-	xml_file_ << "\n<volumeregion name=\"" << name << "\">\n";
+	file_ << "\n<volumeregion name=\"" << name << "\">\n";
 	writeParamMap(*params_);
-	xml_file_ << "</volumeregion>\n";
+	file_ << "</volumeregion>\n";
 	return nullptr;
 }
 
-ImageOutput *XmlExport::createOutput(const char *name) noexcept
+ImageOutput *ExportXml::createOutput(const char *name) noexcept
 {
-	xml_file_ << "\n<output name=\"" << name << "\">\n";
+	file_ << "\n<output name=\"" << name << "\">\n";
 	writeParamMap(*params_);
-	xml_file_ << "</output>\n";
+	file_ << "</output>\n";
 	return nullptr;
 }
 
-RenderView *XmlExport::createRenderView(const char *name) noexcept
+RenderView *ExportXml::createRenderView(const char *name) noexcept
 {
-	xml_file_ << "\n<render_view name=\"" << name << "\">\n";
+	file_ << "\n<render_view name=\"" << name << "\">\n";
 	writeParamMap(*params_);
-	xml_file_ << "</render_view>\n";
+	file_ << "</render_view>\n";
 	return nullptr;
 }
 
-Image *XmlExport::createImage(const char *name) noexcept
+Image *ExportXml::createImage(const char *name) noexcept
 {
-	xml_file_ << "\n<image name=\"" << name << "\">\n";
+	file_ << "\n<image name=\"" << name << "\">\n";
 	writeParamMap(*params_);
-	xml_file_ << "</image>\n";
+	file_ << "</image>\n";
 	return nullptr;
 }
 
-Object *XmlExport::createObject(const char *name) noexcept
+Object *ExportXml::createObject(const char *name) noexcept
 {
 	n_uvs_ = 0;
-	xml_file_ << "\n<object>\n";
-	xml_file_ << "\t<object_parameters name=\"" << name << "\">\n";
+	file_ << "\n<object>\n";
+	file_ << "\t<object_parameters name=\"" << name << "\">\n";
 	writeParamMap(*params_, 2);
-	xml_file_ << "\t</object_parameters>\n";
+	file_ << "\t</object_parameters>\n";
 	++next_obj_;
 	return nullptr;
 }
 
-void XmlExport::setupRender() noexcept
+void ExportXml::setupRender() noexcept
 {
-	xml_file_ << "\n<render>\n";
+	file_ << "\n<render>\n";
 	writeParamMap(*params_);
-	xml_file_ << "</render>\n";
+	file_ << "</render>\n";
 }
 
-void XmlExport::render(std::shared_ptr<ProgressBar> progress_bar) noexcept
+void ExportXml::render(std::shared_ptr<ProgressBar> progress_bar) noexcept
 {
-	xml_file_ << "</scene>\n";
-	xml_file_.flush();
-	xml_file_.close();
+	file_ << "</scene>\n";
+	file_.flush();
+	file_.close();
 }
 
-void XmlExport::setXmlColorSpace(std::string color_space_string, float gamma_val) noexcept
+void ExportXml::setColorSpace(std::string color_space_string, float gamma_val) noexcept
 {
-	xml_color_space_ = Rgb::colorSpaceFromName(color_space_string, ColorSpace::Srgb);
-	xml_gamma_ = gamma_val;
+	color_space_ = Rgb::colorSpaceFromName(color_space_string, ColorSpace::Srgb);
+	gamma_ = gamma_val;
 }
 
 END_YAFARAY